ensure tulip_perspective has not failed before playing cnee recorded events
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@10077 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ unit_test/gui/tulip_run_test.sh
@@ -62,6 +62,12 @@ fi
 # run the test
 sh tulip_replay.sh $TEST_NAME.xns $TLP_INPUT_FILE
 
+# check for a crash
+if [ ! -f tulip.pid ]; then
+  echo "$TEST_NAME test failed: tulip_perspective execution aborted"
+  exit
+fi
+
 # move data/test_gui.tlp if it exist
 if [ -f data/test_gui.tlp ]; then
     mv data/test_gui.tlp .
